requirements.txt
git+https://github.com/huggingface/diffusers.git

huggingface-hub
spaces
transformers
accelerate
safetensors
sentencepiece
dashscope
kernels
torchvision
peft
torchao==0.14.1 
pydantic
pandas
modal

para-attn
lpips
https://huggingface.co/spaces/eleazhong/Qwen-Image-Edit-Angles/resolve/main/bin/sageattention-2.2.0-cp310-cp310-linux_x86_64.whl